refactor!: apollo-client to urql
- also updated graphql-code-generator for urql
