+++ ionic/package.json
@@ -12,18 +12,18 @@
     "generate": "graphql-codegen --config codegen.yml"
   },
   "dependencies": {
-    "@apollo/react-hooks": "^3.1.5",
     "@capacitor/core": "2.1.0",
     "@ionic/react": "^5.0.7",
     "@ionic/react-router": "^5.0.7",
-    "apollo-boost": "^0.4.9",
     "graphql": "^15.0.0",
+    "graphql-tag": "^2.11.0",
     "ionicons": "^5.0.0",
     "react": "^16.13.0",
     "react-dom": "^16.13.0",
     "react-hook-form": "^6.0.0",
     "react-router": "^5.1.2",
-    "react-router-dom": "^5.1.2"
+    "react-router-dom": "^5.1.2",
+    "urql": "^1.10.0"
   },
   "devDependencies": {
     "@capacitor/cli": "2.1.0",
@@ -32,7 +32,7 @@
     "@graphql-codegen/near-operation-file-preset": "^1.17.11",
     "@graphql-codegen/typescript": "1.17.8",
     "@graphql-codegen/typescript-operations": "1.17.8",
-    "@graphql-codegen/typescript-react-apollo": "2.0.6",
+    "@graphql-codegen/typescript-urql": "^2.0.1",
     "@ionic/cli": "^6.9.2",
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.4.0",
